requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Django==4.2.3
django-cors-headers==4.2.0
idna==3.4
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
